+++ projects/mk-kit/src/lib/components/datetime.barrel.ts
@@ -9,3 +9,4 @@ export * from './date-picker';
 export * from './time-picker';
 export * from './date-range-picker';
 export * from './month-picker';
+export * from './week-picker';